problems faced
VIdeo does not stop on changing slide
when returning to a slide, it does not appear on the screen
No file found is not handled by any media handlers except audio, which
crashes.

Work to be Done:
Fix problems above

+++ SmartTrolley/src/imagehandler/SlideImage.java
@@ -68,7 +68,7 @@ public class SlideImage extends ImageView{
 	*@param duration
 	*<p> Date Modified: May 3, 2014
 	*/
-	protected void setDuration(int seconds) {
+	protected void setDuration(double seconds) {
 		
 		duration.setDuration(seconds);
